requirements.txt
amqp==2.2.2
billiard==3.5.0.3
boto3==1.4.8
botocore==1.8.6
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-anymail==1.2
django-appconf==1.0.2
django-avatar==4.1.0
django-cors-middleware==1.3.1
django-crum==0.7.2
django-debug-toolbar==1.9.1
django-facebook==6.0.3
django-filter==1.1.0
django-mailgun==0.9.1
django-oauth-toolkit==1.0.0
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
django-storages==1.6.5
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
docutils==0.14
drf-tracking==1.3.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
Markdown==2.6.9
MarkupSafe==1.0
numpy==1.13.3
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
opencv-python==3.3.1.11
Pillow==4.3.0
psycopg2==2.7.3.2
PyJWT==1.5.3
pytesseract==0.1.7
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.12
simplejson==3.12.0
six==1.11.0
sqlparse==0.2.4
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4